deleted pages folder and moved eventList to /components/events and moved the store folder to the src folder, also cahnged the imports.
