+++ src/components/events/eventsSlice.js
@@ -1,4 +1,4 @@
-import api from "../../../store/api";
+import api from "../../store/api";
 
 const eventApi = api.injectEndpoints({
   endpoints: (build) => ({
